+++ source/renderer/app/actions/sidebar-actions.js
@@ -11,4 +11,5 @@ export default class SidebarActions {
     showSubMenu?: boolean,
   }> = new Action();
   walletSelected: Action<{ walletId: string }> = new Action();
+  hardwareWalletSelected: Action<{ walletId: string }> = new Action();
 }