+++ source/renderer/app/config/sidebarConfig.js
@@ -1,6 +1,7 @@
 // @flow
 import { ROUTES } from '../routes-config';
 import walletsIcon from '../assets/images/sidebar/wallet-ic.inline.svg';
+import hardwareWalletsIcon from '../assets/images/sidebar/hardware-wallet-ic.inline.svg';
 import settingsIcon from '../assets/images/sidebar/settings-ic.inline.svg';
 import paperWalletCertificateIcon from '../assets/images/sidebar/paper-certificate-ic.inline.svg';
 import delegationIcon from '../assets/images/sidebar/delegation-ic.inline.svg';
@@ -18,6 +19,11 @@ export const CATEGORIES_BY_NAME = {
     name: 'WALLETS',
     icon: walletsIcon,
     route: ROUTES.WALLETS.ROOT,
+  },
+  HARDWARE_WALLETS: {
+    name: 'HARDWARE_WALLETS',
+    icon: hardwareWalletsIcon,
+    route: ROUTES.HARDWARE_WALLETS.ROOT,
   },
   PAPER_WALLET_CREATE_CERTIFICATE: {
     name: 'PAPER_WALLET_CREATE_CERTIFICATE',
@@ -80,6 +86,12 @@ export const CATEGORIES_WITH_STAKING = [
   CATEGORIES_BY_NAME.NETWORK_INFO,
 ];
 
+export const CATEGORIES_WITH_HARDWARE_WALLETS = [
+  CATEGORIES_BY_NAME.WALLETS,
+  CATEGORIES_BY_NAME.HARDWARE_WALLETS,
+  CATEGORIES_BY_NAME.SETTINGS,
+];
+
 export const sidebarConfig = {
   CATEGORIES_BY_NAME,
   CATEGORIES,
@@ -87,4 +99,5 @@ export const sidebarConfig = {
   CATEGORIES_WITH_DELEGATION_COUNTDOWN,
   CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN,
   CATEGORIES_WITH_STAKING,
+  CATEGORIES_WITH_HARDWARE_WALLETS,
 };